@@ -50,41 +50,42 @@
   },
   "homepage": "https://github.com/getsaf/shallow-render#readme",
   "peerDependencies": {
-    "@angular/common": "9.x",
-    "@angular/compiler": "9.x",
-    "@angular/core": "9.x",
-    "@angular/forms": "9.x",
-    "@angular/platform-browser": "9.x"
-  },
-  "dependencies": {
-    "ng-mocks": "^9.0.0"
+    "@angular/common": ">=9",
+    "@angular/compiler": ">=9",
+    "@angular/core": ">=9",
+    "@angular/forms": ">=9",
+    "@angular/platform-browser": ">=9"
   },
   "devDependencies": {
-    "@angular/common": "9.x",
-    "@angular/compiler": "9.x",
-    "@angular/core": "9.x",
-    "@angular/forms": "9.x",
-    "@angular/platform-browser": "9.x",
-    "@angular/platform-browser-dynamic": "9.x",
-    "@angular/router": "9.x",
-    "@types/jasmine": "^3.5.4",
+    "@angular/common": "10.x",
+    "@angular/compiler": "10.x",
+    "@angular/core": "10.x",
+    "@angular/forms": "10.x",
+    "@angular/platform-browser": "10.x",
+    "@angular/platform-browser-dynamic": "10.x",
+    "@angular/router": "10.x",
+    "@types/jasmine": "^3.5.11",
+    "@types/jasminewd2": "^2.0.8",
     "@types/node": "^13.7.1",
-    "codelyzer": "^5.2.1",
+    "codelyzer": "^6.0.0-next.1",
     "core-js": "^2.5.7",
     "jasmine": "^3.5.0",
+    "jasmine-core": "^3.5.0",
+    "jasmine-spec-reporter": "^5.0.2",
     "jasmine-ts": "^0.3.0",
-    "karma": "^4.4.1",
+    "karma": "^5.0.0",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-jasmine": "^3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-typescript": "^4.0.0",
-    "prettier": "^1.19.1",
-    "rxjs": "^6.5.4",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.20.1",
+    "karma-typescript": "^5.0.3",
+    "ng-packagr": "^10.0.0",
+    "prettier": "^2.0.5",
+    "rxjs": "^6.5.5",
+    "ts-node": "^8.3.0",
+    "tslint": "^6.1.0",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "^3.7.5",
-    "zone.js": "^0.10.2"
+    "typescript": "^3.9.5",
+    "zone.js": "^0.10.3"
   }
 }
