@@ -15,21 +15,21 @@ module.exports = (config: any) => {
       'node_modules/zone.js/dist/async-test.js',
       'node_modules/zone.js/dist/fake-async-test.js',
       'karma-test-shim.ts',
-      { pattern: 'lib/**/*.ts' }
+      { pattern: 'lib/**/*.ts' },
     ],
     frameworks: ['jasmine', 'karma-typescript'],
     logLevel: config.LOG_INFO,
     port: 9876,
     preprocessors: {
-      '**/*.ts': ['karma-typescript']
+      '**/*.ts': ['karma-typescript'],
     },
     reporters: ['dots', 'karma-typescript', 'kjhtml'],
     singleRun: true,
 
     karmaTypescriptConfig: {
       compilerOptions: {
-        target: 'es2015'
-      }
-    }
+        target: 'es2015',
+      },
+    },
   });
 };
