@@ -10,18 +10,14 @@ interface MyEvent {
 
 @Component({
   selector: 'inner-component',
-  template: `
-    <p>Hello</p>
-  `
+  template: '<p>Hello</p>',
 })
 class InnerComponent {
   @Output() output = new EventEmitter<MyEvent>();
 }
 @Component({
   selector: 'outer-component',
-  template: `
-    <inner-component (output)="outputHandler($event)"></inner-component>
-  `
+  template: '<inner-component (output)="outputHandler($event)"></inner-component>',
 })
 class OuterComponent {
   outputHandler(_event: MyEvent) {
@@ -30,7 +26,7 @@ class OuterComponent {
 }
 
 @NgModule({
-  declarations: [InnerComponent, OuterComponent]
+  declarations: [InnerComponent, OuterComponent],
 })
 class CustomInputModule {}
 //////////////////////////
