@@ -4,7 +4,7 @@ import { Shallow } from '../shallow';
 ////// Module Setup //////
 @Component({
   selector: 'list-container',
-  template: '<ul><ng-content *ngIf="!collapsed"></ng-content></ul>'
+  template: '<ul><ng-content *ngIf="!collapsed"></ng-content></ul>',
 })
 class ListContainerComponent {
   protected collapsed = false;
@@ -19,7 +19,7 @@ class ListContainerComponent {
 
 @Component({
   selector: 'list-item',
-  template: '<li [class.bold]="bold"><ng-content></ng-content></li>'
+  template: '<li [class.bold]="bold"><ng-content></ng-content></li>',
 })
 class ListItemComponent {
   @Input() bold = false;
@@ -31,12 +31,12 @@ class ListItemComponent {
     <list-container #container>
       <list-item (click)="container.collapse()">Chuck Norris</list-item>
     </list-container>
-  `
+  `,
 })
 class AwesomeListComponent {}
 
 @NgModule({
-  declarations: [ListContainerComponent, ListItemComponent, AwesomeListComponent]
+  declarations: [ListContainerComponent, ListItemComponent, AwesomeListComponent],
 })
 class ListModule {}
 //////////////////////////
