@@ -1,34 +1,35 @@
 import { Component, DebugElement, Directive, EventEmitter, Input, Output, Type, InjectionToken } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { MockComponent, MockDirective } from 'ng-mocks';
 import { Rendering } from './rendering';
 import { TestSetup } from './test-setup';
+import { mockComponent } from '../tools/mock-component';
+import { mockDirective } from '../tools/mock-directive';
 
 @Component({
-  template: '<outer></outer>'
+  template: '<outer></outer>',
 })
 class TestHostComponent {}
 
 @Directive({
-  selector: '[directive-to-mock]'
+  selector: '[directive-to-mock]',
 })
 class WillBeMockedDirective {}
 
 @Directive({
-  selector: '[structuralDirectiveToMock]'
+  selector: '[structuralDirectiveToMock]',
 })
 class WillBeMockedStructuralDirective {
   @Input() structuralDirectiveToMock!: string;
 }
 
 @Directive({
-  selector: '[test-directive]'
+  selector: '[test-directive]',
 })
 class TestDirective {}
 
 @Directive({
-  selector: '[other-directive]'
+  selector: '[other-directive]',
 })
 class OtherDirective {
   @Input('other-directive') otherDirective!: string;
@@ -53,7 +54,7 @@ class OtherDirective {
         </div>
       </div>
     </div>
-  `
+  `,
 })
 class OuterComponent {
   @Output() markedAsOutput = new EventEmitter<string>();
@@ -62,19 +63,19 @@ class OuterComponent {
 
 @Component({
   selector: 'inner',
-  template: '<span test-directive>sub</span>'
+  template: '<span test-directive>sub</span>',
 })
 class InnerComponent {}
 
 @Component({
   selector: 'component-to-mock',
-  template: '<span>this will not render<span>'
+  template: '<span>this will not render<span>',
 })
 class WillBeMockedComponent {}
 
 @Component({
   selector: 'other',
-  template: '<span>other</span>'
+  template: '<span>other</span>',
 })
 class OtherComponent {}
 const MY_TOKEN = new InjectionToken<boolean>('My boolean token');
@@ -89,9 +90,9 @@ describe('Rendering', () => {
   let MockedStructuralDirective: Type<WillBeMockedStructuralDirective>;
 
   beforeEach(async () => {
-    MockedDirective = MockDirective(WillBeMockedDirective);
-    MockedStructuralDirective = MockDirective(WillBeMockedStructuralDirective);
-    MockedComponent = MockComponent(WillBeMockedComponent);
+    MockedDirective = mockDirective(WillBeMockedDirective);
+    MockedStructuralDirective = mockDirective(WillBeMockedStructuralDirective);
+    MockedComponent = mockComponent(WillBeMockedComponent);
     testSetup = new TestSetup(OuterComponent, class {});
     testSetup.mockCache.add(WillBeMockedDirective, MockedDirective);
     testSetup.mockCache.add(WillBeMockedComponent, MockedComponent);
@@ -106,8 +107,8 @@ describe('Rendering', () => {
         OtherDirective,
         MockedComponent,
         MockedDirective,
-        MockedStructuralDirective
-      ]
+        MockedStructuralDirective,
+      ],
     })
       .compileComponents()
       .then(() => {
@@ -303,7 +304,7 @@ describe('Rendering', () => {
         testSetup
       );
       const found = findStructuralDirective(WillBeMockedStructuralDirective, {
-        query: d => d.structuralDirectiveToMock === 'second-one'
+        query: d => d.structuralDirectiveToMock === 'second-one',
       });
       renderStructuralDirective(found);
 
@@ -367,7 +368,7 @@ describe('Rendering', () => {
       const { findStructuralDirective } = new Rendering(fixture, element, instance, {}, testSetup);
 
       const found = findStructuralDirective(WillBeMockedStructuralDirective, {
-        query: d => d.structuralDirectiveToMock === 'second-one'
+        query: d => d.structuralDirectiveToMock === 'second-one',
       });
 
       expect(found.structuralDirectiveToMock).toBe('second-one');
