@@ -35,37 +35,37 @@ const jasmineMatchers = {
   toHaveFound: () => ({
     compare: (actual: QueryMatch<object>, expected: number) => ({
       pass: actual.length === expected,
-      message: `Expected to find exactly ${expected} but found ${actual.length}`
-    })
+      message: `Expected to find exactly ${expected} but found ${actual.length}`,
+    }),
   }),
 
   toHaveFoundOne: () => ({
     compare: (actual: QueryMatch<object>) => ({
       pass: actual.length === 1,
-      message: `Expected to find exactly 1 but found ${actual.length}`
-    })
+      message: `Expected to find exactly 1 but found ${actual.length}`,
+    }),
   }),
 
   toHaveFoundMoreThan: () => ({
     compare: (actual: QueryMatch<object>, expected: number) => ({
       pass: actual.length > expected,
-      message: `Expected to find more than ${expected} but found ${actual.length}`
-    })
+      message: `Expected to find more than ${expected} but found ${actual.length}`,
+    }),
   }),
 
   toHaveFoundLessThan: () => ({
     compare: (actual: QueryMatch<object>, expected: number) => ({
       pass: actual.length < expected,
-      message: `Expected to find less than ${expected} but found ${actual.length}`
-    })
-  })
+      message: `Expected to find less than ${expected} but found ${actual.length}`,
+    }),
+  }),
 };
 
 const jestMatchers = {
   toHaveFound: jasmineToJestMatcher(jasmineMatchers.toHaveFound),
   toHaveFoundLessThan: jasmineToJestMatcher(jasmineMatchers.toHaveFoundLessThan),
   toHaveFoundMoreThan: jasmineToJestMatcher(jasmineMatchers.toHaveFoundMoreThan),
-  toHaveFoundOne: jasmineToJestMatcher(jasmineMatchers.toHaveFoundOne)
+  toHaveFoundOne: jasmineToJestMatcher(jasmineMatchers.toHaveFoundOne),
 };
 
 /////////////////////
