@@ -1,13 +1,14 @@
-import { forwardRef, PipeTransform, Type } from '@angular/core';
-import { TestBed } from '@angular/core/testing';
-import { MockDeclaration, MockPipe } from 'ng-mocks';
+import { PipeTransform, Type } from '@angular/core';
 import { AngularModule } from '../models/angular-module';
 import { TestSetup } from '../models/test-setup';
-import { testFramework } from '../test-frameworks/test-framework';
 import { mockModule } from './mock-module';
 import { directiveResolver, ngModuleResolver } from './reflect';
-import { isModuleWithProviders, isPipeTransform, isClass } from './type-checkers';
+import { isModuleWithProviders, isPipeTransform, isClass, declarationType } from './type-checkers';
 import { CustomError } from '../models/custom-error';
+import { mockPipe } from './mock-pipe';
+import { mockDirective } from './mock-directive';
+import { mockComponent } from './mock-component';
+import { TestBed } from '@angular/core/testing';
 
 export type NgMockable = AngularModule | Type<any> | Type<PipeTransform> | any[];
 
@@ -31,23 +32,19 @@ export function ngMock<TThing extends NgMockable | NgMockable[]>(thing: TThing, 
     if (ngModuleResolver.isNgModule(thing) || isModuleWithProviders(thing)) {
       mock = mockModule(thing, setup);
     } else if (isPipeTransform(thing)) {
-      mock = MockPipe(thing, setup.mockPipes.get(thing));
+      mock = mockPipe(thing, setup.mockPipes.get(thing));
     } else if (isClass(thing)) {
-      mock = MockDeclaration(thing);
-      fixEmptySelector(thing, mock);
-      if (shouldRenderOnInit(setup, thing)) {
-        renderTemplateOnInit(mock);
-      }
       const stubs = setup.mocks.get(thing);
-      if (stubs) {
-        mock = extendMockWithStubs(mock, stubs);
-      }
-      Object.defineProperty(mock, 'name', { value: `MockOf${thing.name}` });
-      // Provide our mock in place of any other usage of 'thing'.
-      // This makes `ViewChild` and `ContentChildren` selectors work!
-      TestBed[isComponent(mock) ? 'overrideComponent' : 'overrideDirective'](mock, {
-        add: { providers: [{ provide: thing, useExisting: forwardRef(() => mock) }] }
-      });
+      mock =
+        declarationType(thing) === 'Component'
+          ? mockComponent(thing, { stubs })
+          : mockDirective(thing, {
+              stubs,
+              renderContentsOnInit:
+                setup.withStructuralDirectives.get(thing) ||
+                (setup.alwaysRenderStructuralDirectives && setup.withStructuralDirectives.get(thing) !== false),
+            });
+      fixEmptySelector(thing, mock);
     } else {
       throw new DoNotKnowHowToMockError(thing);
     }
@@ -66,56 +63,18 @@ class DoNotKnowHowToMockError extends CustomError {
 class MockError extends CustomError {
   constructor(thing: any, error: any) {
     super(
-      `Shallow ran into some trouble mocking ${thing?.name ||
-        thing}. Try skipping it with dontMock or neverMock.\n------------- MOCK ERROR -------------\n${error?.stack ||
-        error}\n----------- END MOCK ERROR -----------`
+      `Shallow ran into some trouble mocking ${
+        thing?.name || thing
+      }. Try skipping it with dontMock or neverMock.\n------------- MOCK ERROR -------------\n${
+        error?.stack || error
+      }\n----------- END MOCK ERROR -----------`
     );
   }
 }
 
-const extendMockWithStubs = (mock: Type<any>, stubs: object): Type<any> => {
-  class MockWithStubs extends mock {
-    constructor() {
-      super();
-      Object.assign(this, stubs);
-      Object.keys(stubs).forEach(key => {
-        if (typeof this[key] === 'function') {
-          testFramework.spyOn(this, key);
-        }
-      });
-    }
-  }
-  return MockWithStubs;
-};
-
-const shouldRenderOnInit = (setup: TestSetup<any>, thing: any) =>
-  setup.withStructuralDirectives.get(thing as Type<any>) ||
-  (setup.alwaysRenderStructuralDirectives && setup.withStructuralDirectives.get(thing as Type<any>) !== false);
-
-const isComponent = (thing: Type<any>) => {
-  if (directiveResolver.isDirective(thing)) {
-    const metadata = directiveResolver.resolve(thing);
-    return 'template' in metadata || 'templateUrl' in metadata;
-  }
-  return false;
-};
-
 const fixEmptySelector = (thing: Type<any>, mock: Type<any>) => {
-  const annotations = directiveResolver.resolve(thing);
-  if (!annotations.selector) {
+  const { selector } = directiveResolver.resolve(thing);
+  if (!selector) {
     TestBed.overrideDirective(mock, { add: { selector: `.__${mock.name}-selector` } });
   }
-  return thing;
-};
-
-const renderTemplateOnInit = (mock: Type<any>) => {
-  const originalInit = (mock.prototype.ngOnInit = mock.prototype.ngOnInit ?? (() => {}));
-  testFramework.spyOn(mock.prototype, 'ngOnInit', function() {
-    try {
-      // @ts-ignore
-      originalInit.call(this);
-      // @ts-ignore
-      this.__render();
-    } catch (e) {}
-  });
 };
